+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - `wahl.ini` now comes with \r\n line endings to support editing with notepad on Windows
 - ballot written with Windows line endings to support editing with notepad on Windows
+- improve formatting of (empty) ballot
 
 ## [0.1.1] - 2017-11-16